# improve pug_escape performance