This improves the performance of `pug_escape()` by replacing the Regular Expression that searches for chars that need to be escaped with calls to`indexOf()`.

I wrote a benchmark that tests the performance with strings containing chars that need to be escaped at different positions:

```
new - nothing to escape x 3,636,837 ops/sec ±0.42% (89 runs sampled)
old - nothing to escape x 1,186,265 ops/sec ±1.33% (91 runs sampled)
new - chars to escape - at end x 2,183,489 ops/sec ±0.63% (90 runs sampled)
old - chars to escape - at end x 851,802 ops/sec ±0.35% (93 runs sampled)
new - chars to escape - at middle x 303,275 ops/sec ±0.38% (91 runs sampled)
old - chars to escape - at middle x 246,988 ops/sec ±0.84% (89 runs sampled)
new - chars to escape - at start x 153,486 ops/sec ±0.70% (91 runs sampled)
old - chars to escape - at start x 138,075 ops/sec ±0.37% (90 runs sampled)
```

The code of the benchmark can be found here:
https://github.com/SerayaEryn/pug/blob/improve-pug_escape-performance-benchmark/packages/pug-runtime/pugEscapeBenchmark.js