@SerayaEryn 2019-02-09
I noticed that the escpae function from jade is faster if there are chars that need to be escaped in the first half of the string.
I added some more optimization for this case.

```
new - nothing to escape x 3,715,290 ops/sec ±2.93% (88 runs sampled)
old - nothing to escape x 1,197,092 ops/sec ±0.65% (91 runs sampled)
new - chars to escape - at end x 2,173,065 ops/sec ±0.59% (89 runs sampled)
old - chars to escape - at end x 865,124 ops/sec ±0.34% (92 runs sampled)
new - chars to escape - at middle x 310,349 ops/sec ±0.73% (88 runs sampled)
old - chars to escape - at middle x 248,706 ops/sec ±0.44% (87 runs sampled)
new - chars to escape - at start x 455,263 ops/sec ±0.35% (92 runs sampled)
old - chars to escape - at start x 139,149 ops/sec ±0.28% (90 runs sampled)
```